feat: add action field to client registration form
Fixes a small inconsistency in the HTML code where 2 different inputs had the same name and id.

Removes one of the submit buttons, since having 2 of them was causing a bug when submitting the form data and places their code inside the <form> tag.

Adds a different value to form's action attribute.